Bug 37412: Style placeholder text in the OPAC
This patch adds a custom style to input placeholder text so that it is
visibly different from the style of text entered in the input.

To test, apply the patch and rebuild the OPAC CSS
(https://wiki.koha-community.org/wiki/Working_with_SCSS_in_the_OPAC_and_staff_interface)

Go to the OPAC and check the appearance of the placeholder text in the
main search bar. Sign off if the vibes are right.

Sponsored-by: Athens County Public Libraries
Signed-off-by: Roman Dolny <roman.dolny@jezuici.pl>
Signed-off-by: Paul Derscheid <paul.derscheid@lmscloud.de>
Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>