Bug 37333: (follow-up) Update existing tests
We add a level of bracket nesting. I checked with Nick that this
shouldn't cause any issues in the queries and we agreed it made sense so
I've gone ahead and updated the tests to reflect the change.

Signed-off-by: Martin Renvoize <martin.renvoize@ptfs-europe.com>
(cherry picked from commit 21dc76f35b54adac2fed03d0919ba08fdeec948a)
Signed-off-by: Lucas Gass <lucas@bywatersolutions.com>