Bug 35044: (QA follow-up): Fix flaky tests
To test, run the following test file a couple times until it fails:
prove t/db_dependent/Koha/Object/Mixin/AdditionalFields.t

My testing this failed around 50% of the time so it should not take
long to reproduce. I believe this is because the order of the array
elements may differ.

Apply this patch and run the test file again many times. It should
never fail.

Signed-off-by: Katrin Fischer <katrin.fischer@bsz-bw.de>